HOTFIX Support iOS 26
